refactor: change name style
