+++ src/helpers/responseCodes.js
@@ -1,0 +1,7 @@
+const RESPONSE_CODES = {
+  OK: 200,
+  Created: 201,
+  Deleted: 204,
+};
+
+module.exports = { RESPONSE_CODES };
